fix: resolve 492 pre-existing ruff findings across the repo's Python code
Fixed by 8 parallel review-and-fix passes, partitioned by disjoint file
sets. Dominant categories: subprocess.run missing explicit check= (each
call site reviewed for actual intent — check=False where the caller
already inspects returncode/stdout/stderr, matching existing behavior),
bare except Exception narrowed to specific types or explicitly annotated
with # noqa + reason where this repo's established fail-open hook idiom
applies, shebang/executable-bit mismatches (EXE001/EXE002), and a long
tail of ruff's own safe auto-fixes (SIM102/SIM103/RUF059/C408/UP031/
ISC004/PIE810/RUF015/DTZ/TRY004/B023/SIM115).

8 findings intentionally left (N999 — numbered redteam probe files
dynamically loaded by filename; renaming would break the harness).

Two pre-existing latent gaps surfaced but deliberately not changed
(behavior-preserving check=False kept, not check=True, since that would
be a behavior change beyond a lint fix): mutation_kill_loop.py's
run_scoped_stryker and both mutation_kill_loop variants' git_revert()
have zero downstream inspection of their subprocess exit codes. Worth a
follow-up issue.

Zero regressions: full pytest dir list run after all fixes landed —
4216 passed, 1 skipped, 2 failed (both pre-existing eslint-sandbox
timeout flakes unrelated to this change, not introduced by it).

+++ plugins/dev-team/skills/agent-readiness/scanner.py
@@ -168,7 +168,7 @@ def c4_module_size(root: Path, cfg: dict) -> dict:
     if not counts:
         return _score(0, "no source files detected for size analysis")
     counts.sort()
-    p90 = counts[min(len(counts) - 1, int(round(0.9 * (len(counts) - 1))))]
+    p90 = counts[min(len(counts) - 1, round(0.9 * (len(counts) - 1)))]
     th = cfg["thresholds"]
     if p90 <= th["file_size_p90_small"]:
         return _score(
